backend/pyproject.toml
[project]
name = "aetheris-backend"
version = "0.1.0"
description = "Session kernel and world engine for AETHERIS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "fastapi>=0.115.12,<1.0.0",
    "pydantic-settings>=2.9.1,<3.0.0",
    "prometheus-client>=0.22.1,<1.0.0",
    "PyJWT>=2.10.1,<3.0.0",
    "redis>=5.2.1,<6.0.0",
    "sqlalchemy>=2.0.40,<3.0.0",
    "uvicorn[standard]>=0.34.2,<1.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1,<0.29.0",
    "mypy>=1.15.0,<2.0.0",
    "pytest>=8.3.5,<9.0.0",
    "ruff>=0.11.6,<0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
packages = ["app"]
